mejora sectorservice

+++ Decimatio.Domain/Interfaces/ISectorRepository.cs
@@ -7,6 +7,6 @@
         Task<Sector> GetById(long idSector);
         Task<int> AddSector(Sector sector);
         Task<bool> UpdateSector(Sector sector);
-        Task<bool> DeleteSector(int idSector);
+        Task<bool> DeleteSector(long idSector);
     }
 }
